+++ android/app/src/main/java/com/fridgechef/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.reactnativetemplate
+package com.fridgechef
 
 import android.app.Application
 import com.facebook.react.PackageList
